requirements.txt
django-scheduler
bcrypt==4.0.1
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.14
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-environ==0.9.0
django-import-export==2.8.0
et-xmlfile==1.1.0
fpdf==1.7.2
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
tablib==3.2.1
typing-extensions==4.3.0
xlrd==2.0.1
xlwt==1.3.0
requests==2.31.0
urllib3==1.26.*
djangorestframework==3.14.0
